requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
Faker==33.1.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
